% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG.R
\name{addAnnotationDEG}
\alias{addAnnotationDEG}
\title{Annotation of differentially expressed genes}
\usage{
addAnnotationDEG(geneCount, fit, contrast)
}
\arguments{
\item{geneCount}{summarized read counts of genes.}

\item{fit}{output of \code{\link{DEG}} function that contains ranking of differentially expressed genes.}

\item{contrast}{a contrast from contrast matrix, whose ranking is required.}
}
\value{
Annotated ranking of differentially expressed genes of given contrast. The output can be saved using write.csv or \code{\link[openxlsx]{write.xlsx}}.
}
\description{
Add gene annotation to ranked differentially expressed genes for a given contrast, using output of \code{\link{DEG}} function.
}
