% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEJ.R
\name{addAnnotationDEJ}
\alias{addAnnotationDEJ}
\title{Annotation of differentially expressed junctions}
\usage{
addAnnotationDEJ(JunctionMatrixA, fit, contrast)
}
\arguments{
\item{JunctionMatrixA}{annotated junction matrix containing junction read counts, produced by \code{\link{JunctionMatrixAnnotation}}.}

\item{fit}{output of \code{\link{DEJ}}, that contains ranking of differentially expressed junctions.}

\item{contrast}{contrast from contrast matrix, whose ranking is required.}
}
\value{
Annotated ranking of differentially expressed junctions of a given contrast. The output can be saved using write.csv or \code{\link[openxlsx]{write.xlsx}}.
}
\description{
Add annotation to ranked differentially expressed junctions of given contrast returned by \code{\link{DEJ}}.
}
