% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addAnnotationRnaSeq.R
\name{addAnnotationRnaSeq}
\alias{addAnnotationRnaSeq}
\title{Add annotation to EP/IP events}
\usage{
addAnnotationRnaSeq(fit, annotation = annotation)
}
\arguments{
\item{fit}{output of \code{\link{EPrnaseq}}/\code{\link{iPrnaseq}}.}

\item{annotation}{matrix; contains annotation of exons and introns, created by \code{\link{readMembershipMatrix}}.}
}
\value{
Annotated matrix of ranked cassette exon/intron retention events. Output of addAnnotationRnaSeq can be passed to \code{\link{getPvaluesByContrast}} to find differentially expressed intron retention events in a given contrast.
}
\description{
Adds the associated information for each ranked cassette exon/intron retention event generated by \code{\link{EPrnaseq}}/\code{\link{iPrnaseq}}. The information includes location of the event (chromosome, start, stop, strand), position in genome and the associated gene name.
}
