% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connected.exon.R
\name{connected.exons.igraph}
\alias{connected.exons.igraph}
\title{Connected Exons using igraph}
\usage{
connected.exons.igraph(
  cmm = cmm,
  event = event,
  ls_exon = ls_exon,
  rs_exon = rs_exon,
  s.exon = NULL,
  eventtype = NULL
)
}
\arguments{
\item{cmm}{}

\item{event}{}

\item{ls_exon}{}

\item{rs_exon}{}

\item{s.exon}{}

\item{eventtype}{}
}
\value{

}
\description{
Connected Exons using igraph
}
