% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countMatrixGenes.R
\name{countMatrixGenes}
\alias{countMatrixGenes}
\title{Count Matrix Genes}
\usage{
countMatrixGenes(
  JunctionMatrix,
  annotation,
  intronList = c(intron_A, intron_B, intron_C),
  exonList = c(out_A, out_B, out_C)
)
}
\arguments{
\item{JunctionMatrix}{matrix of junction read counts times samples, generated by \code{\link{getJunctionCountMatrix}}.}

\item{annotation}{gene features and meta-featues annotation file generated by \code{\link{readMembershipMatrix}}, saved by default as Annotation.Rdata.}

\item{intronList}{intron read counts per gene generated by Rsubread package and saved in counts_introns.Rdata file, as per preprocessing instructions.}

\item{exonList}{exon read counts per gene generated by Rsubread package and saved in counts_exons.Rdata file, as per preprocessing instructions.}
}
\value{
Gcount list contains gene-wise read count summarization of meta-features times samples in the study. The output is saved as Gcount.Rdata.
}
\description{
This function creates association of metafeatures such as exons, introns and junctions times sample for each gene. It requires a junction matrix, annotation matrix (generated by default using \code{\link{readMembershipMatrix}}) and summarized exon and intron read counts. It should be run only after running \code{\link{readMembershipMatrix}} and \code{\link{intronMembershipMatrix}}.
}
