% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG.R
\name{cpmCountsDEG}
\alias{cpmCountsDEG}
\title{cpmCountsDEG}
\usage{
cpmCountsDEG(
  geneCount,
  filename,
  designM = designM,
  contrastM = contrastM,
  Groups = Groups
)
}
\arguments{
\item{geneCount}{summarized read counts of genes.}

\item{filename}{filename in which output of addAnnotationDEG is saved.}

\item{designM}{design matrix required by limma}

\item{contrastM}{contrast matrix required by limma.}

\item{Groups}{list of sample groups. \cr
Example: If there are two sample groups with three samples each, 'Groups' should be formed as:
\enumerate{
\item numeric: c(1, 1, 1, 2, 2, 2)
}}
}
\value{
Read counts and log2cpm expression of given contrast for ranked differentially expressed genes.
}
\description{
Generates read counts and log2cpm expression for differentially expressed genes for a given contrast, using output of \code{\link{addAnnotationDEG}} function.
}
\references{
\enumerate{
\item Robinson, M. D., McCarthy, D. J. & Smyth, G. K. edgeR: A Bioconductor package for differential expression analysis of digital gene expression data. Bioinformatics 26, 139–140 (2009)
}
}
