% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEJ.R
\name{cpmCountsDEJ}
\alias{cpmCountsDEJ}
\title{cpmCountsDEJ}
\usage{
cpmCountsDEJ(
  JunctionMatrix,
  filename,
  designM = designM,
  contrastM = contrastM,
  Groups = Groups
)
}
\arguments{
\item{JunctionMatrix}{matrix containing read counts for junctions.}

\item{filename}{filename in which output of \code{\link{addAnnotationDEJ}} is saved.}

\item{designM}{design matrix required by limma}

\item{contrastM}{contrast matrix required by limma.}

\item{Groups}{list of sample groups. \cr
Example: If there are two sample groups with three samples each, 'Groups' should be formed as:
\enumerate{
\item numeric: c(1, 1, 1, 2, 2, 2)
}}
}
\value{
Read counts and logcpm expression of given contrast.
}
\description{
Save read counts and log2cpm expression of differentially expressed junctions.
}
\references{
\enumerate{
\item Robinson, M. D., McCarthy, D. J. & Smyth, G. K. edgeR: A Bioconductor package for differential expression analysis of digital gene expression data. Bioinformatics 26, 139–140 (2009)
}
}
