% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpmCountsEP.R
\name{cpmCountsEP}
\alias{cpmCountsEP}
\title{cpmCountsEP}
\usage{
cpmCountsEP(filename, designM = designM, Groups, Gcount = Gcount)
}
\arguments{
\item{filename}{file in which ExonPointer/IntronPointer ranking of a contrast is saved. \cr *(Filename should be in csv format.)}

\item{designM}{design matrix.}

\item{Groups}{list of sample groups. \cr
Example: If there are two sample groups with three samples each, 'Groups' should be formed as: c(1, 1, 1, 2, 2, 2).}

\item{Gcount}{list; contains gene-wise  matrix of meta-features read counts times samples, generated by \code{\link{countMatrixGenes}}.}
}
\value{
Read counts and log2cpm expression of ranked cassette exon/intron retention events for the given contrast.
}
\description{
This function requires ranking of cassette exon/intron retention events as generated by \code{\link{getPvaluesByContrast}} to generate raw read counts and log2cpm expression for the ranking of events in a given contrast only.
}
\references{
\enumerate{
\item Robinson, M. D., McCarthy, D. J. & Smyth, G. K. edgeR: A Bioconductor package for differential expression analysis of digital gene expression data. Bioinformatics 26, 139–140 (2009)
 }
}
