% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG.R
\name{DEG}
\alias{DEG}
\title{Differentially Expressed Genes}
\usage{
DEG(geneCount, designM = designM, contrastM = contrastM, Groups = Groups)
}
\arguments{
\item{geneCount}{summarized read counts of genes.}

\item{designM}{design matrix required by limma.}

\item{contrastM}{contrast matrix required by limma.}

\item{Groups}{list of sample groups. \cr
Example: If there are two sample groups with three samples each, 'Groups' should be formed as:
\enumerate{
\item numeric: c(1, 1, 1, 2, 2, 2)
}}
}
\value{
Saves raw gene counts and log2cpm expression for all genes. Meta-data generated through this function is saved in fit2.Rdata file. Further, annotation of this meta-data is performed by \code{\link{addAnnotationDEG}} function. Contrast-wise ranking of annotated differentially expressed genes can be obtained using \code{\link{cpmCountsDEG}} function.
}
\description{
A wrapper function of limma package to find differentially expressed genes, given summarized read counts of genes obtained from \code{\link[Rsubread]{featureCounts}} function or preprocessing intructions.
}
\references{
\enumerate{
\item Robinson, M. D., McCarthy, D. J. & Smyth, G. K. edgeR: A Bioconductor package for differential expression analysis of digital gene expression data. Bioinformatics 26, 139–140 (2009)
}
}
