% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEJ.R
\name{DEJ}
\alias{DEJ}
\title{Differentially Expressed Junctions}
\usage{
DEJ(JunctionMatrix, designM = designM, contrastM = contrastM, Groups = Groups)
}
\arguments{
\item{JunctionMatrix}{matrix containing read counts for junctions, obtained using \code{\link{getJunctionCountMatrix}} .}

\item{designM}{design matrix required by limma}

\item{contrastM}{contrast matrix required by limma.}

\item{Groups}{list of sample groups. \cr
Example: If there are two sample groups with three samples each, 'Groups' should be formed as:
\enumerate{
\item numeric: c(1, 1, 1, 2, 2, 2)
}}
}
\value{
The output is ranked differentially expressed junctions. Meta-data is saved as fit2.Rdata in folderSRA directory. The ranking can be annotated using \code{\link{addAnnotationDEJ}}. The annotated and ranked differentially expressed junctions for a given contrast (as given in contrast matrix) can be saved using \code{\link{cpmCountsDEJ}}
}
\description{
Find differentially expressed junctions, provided a junction matrix as input (generated by \code{\link{getJunctionCountMatrix}}). This function uses standard limma package (\code{\link[limma]{eBayes}}) to find differentially expressed junctions.
}
\references{
\enumerate{
\item Robinson, M. D., McCarthy, D. J. & Smyth, G. K. edgeR: A Bioconductor package for differential expression analysis of digital gene expression data. Bioinformatics 26, 139–140 (2009)
}
}
