% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS.prep.R
\name{.cmm}
\alias{.cmm}
\title{Title}
\usage{
.cmm(expression = expression, RMM = RMM, iMM = iMM)
}
\arguments{
\item{expression}{}

\item{RMM}{}

\item{iMM}{}

\item{geneID}{}
}
\value{

}
\description{
Title
}
