% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getJunctionCountMatrix.R
\name{.correctJunctionCoordinate}
\alias{.correctJunctionCoordinate}
\title{correctJunctionCoordinate}
\usage{
.correctJunctionCoordinate(bed)
}
\value{

}
\description{
Internal function for \code{\link{getJunctionCountMatrix}}, not to be run separately.
}
