% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkAnalysis.R
\name{.epip.exp}
\alias{.epip.exp}
\title{Prepare files for hs and simscore}
\usage{
.epip.exp(ep, ip, ep.exp, ip.exp)
}
\arguments{
\item{ep}{}

\item{ip}{}

\item{ep.exp}{}

\item{exp}{}
}
\value{

}
\description{
Prepare files for hs and simscore
}
