% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPrnaseq.R
\name{.ePrnaseqFunction}
\alias{.ePrnaseqFunction}
\title{EP function}
\usage{
.ePrnaseqFunction(
  counts = y,
  RMM = RMmeber,
  contrastM = contrastM,
  designM = designM,
  Groups = Groups
)
}
\value{

}
\description{
Internal function of \code{\link{EPrnaseq}}, not to be called separately.
}
