% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intronGTFparser.R
\name{.extractIntrons}
\alias{.extractIntrons}
\title{extractIntrons}
\usage{
.extractIntrons(sGTF)
}
\value{

}
\description{
Internal function of \code{\link{intronGTFparser}}, not to be called separately.
}
