% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPrnaseq.R
\name{.fitEPrnaseqModel}
\alias{.fitEPrnaseqModel}
\title{EP model fitting}
\usage{
.fitEPrnaseqModel(gene, counts, RMmeber, contrastM, designM, Groups, threshold)
}
\value{

}
\description{
Internal function of \code{\link{EPrnaseq}}, not to be called separately.
}
