% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iPrnaseq.R
\name{.fitiPrnaseqModel}
\alias{.fitiPrnaseqModel}
\title{fitiPrnaseqModel}
\usage{
.fitiPrnaseqModel(gene, counts, iMMeber, contrastM, designM, Groups, threshold)
}
\value{

}
\description{
Internal function of \code{\link{iPrnaseq}}, not to be called separately.
}
