% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countMatrixGenes.R
\name{.getEIJcounts}
\alias{.getEIJcounts}
\title{.getEIJcounts}
\usage{
.getEIJcounts(Gannotation, GjunctionCount, GintronCount, GexonCount)
}
\value{

}
\description{
Internal function of \code{\link{countMatrixGenes}}, not to be called separately.
}
