% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countMatrixGenes.R
\name{.getNumericCount}
\alias{.getNumericCount}
\title{.getNumericCount}
\usage{
.getNumericCount(counts)
}
\value{

}
\description{
Internal function of \code{\link{countMatrixGenes}}, not to be run separately.
}
