% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMembershipMatrix.R
\name{.Gstructure}
\alias{.Gstructure}
\title{Gene structure}
\usage{
.Gstructure(sGTF, sJM)
}
\value{

}
\description{
Internal function for \code{\link{readMembershipMatrix}}, not to be called separately.
}
