% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iPrnaseq.R
\name{.iPrnaseqFunction}
\alias{.iPrnaseqFunction}
\title{iPrnaseqFunction}
\usage{
.iPrnaseqFunction(
  counts = y,
  iMM = iMMeber,
  contrastM = contrastM,
  designM = designM,
  Groups = Groups
)
}
\value{

}
\description{
Internal function of \code{\link{iPrnaseq}}, not to be called separately.
}
