% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intronMembershipMatrix.R
\name{.iStructure}
\alias{.iStructure}
\title{intron Structure}
\usage{
.iStructure(sGTF)
}
\value{

}
\description{
Internal function for \code{\link{intronMembershipMatrix}}, not to be called separately.
}
