% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareCounts.R
\name{.prepareCounts}
\alias{.prepareCounts}
\title{Prepare counts for ExonPointer and IntronPointer}
\usage{
.prepareCounts(y, designM, Groups = NULL, threshold = NULL, ...)
}
\value{

}
\description{
Internal function for \code{\link{EPrnaseq}}/\code{\link{iPrnaseq}}, not to be run separately.
}
\references{
Henrik Bengtsson (2017). matrixStats: Functions that Apply to Rows and Columns of Matrices (and to Vectors). R package version 0.52.2. https://github.com/HenrikBengtsson/matrixStats
}
