% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propagation.R
\name{.propagation}
\alias{.propagation}
\title{Transcript Structure: seed propagation function}
\usage{
.propagation(
  cmm = cmm,
  ts.p = ts.p,
  cjmm = cjmm,
  se = se,
  s_exon = s_exon,
  event = event,
  eventtype,
  connected.exon = connected.exon,
  s.exon = NULL,
  keep.intron = FALSE
)
}
\arguments{
\item{cmm}{}

\item{ts.p}{}

\item{cjmm}{}

\item{se}{}

\item{s_exon}{}

\item{event}{}

\item{eventtype}{}

\item{connected.exon}{}

\item{s.exon}{}

\item{keep.intron}{}
}
\value{

}
\description{
Transcript Structure: seed propagation function
}
