% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPvaluesByContrast.R
\name{.removeLECounts}
\alias{.removeLECounts}
\title{Remove low-expressed reads}
\usage{
.removeLECounts(y, designM, Groups = NULL, threshold = 6.32, ...)
}
\value{

}
\description{
Internal function of \code{\link{EPrnaseq}}/\code{\link{iPrnaseq}}, not to be called separately.
}
\references{
Henrik Bengtsson (2017). matrixStats: Functions that Apply to Rows and Columns of Matrices (and to Vectors). R package version 0.52.2. https://github.com/HenrikBengtsson/matrixStats
}
