% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS.prep.R
\name{.removeLECountsTS}
\alias{.removeLECountsTS}
\title{Title}
\usage{
.removeLECountsTS(
  Gcount = Gcount,
  designM = designM,
  Groups = Groups,
  threshold = 6.32,
  ...
)
}
\arguments{
\item{Gcount}{}

\item{designM}{}

\item{Groups}{}

\item{threshold}{}

\item{...}{}
}
\value{

}
\description{
Title
}
