% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seed.exon.R
\name{.s.exon.ip.incl}
\alias{.s.exon.ip.incl}
\title{Title}
\usage{
.s.exon.ip.incl(ip_event = ip_event, cmm = cmm, annotation = annotation)
}
\arguments{
\item{ip_event}{}

\item{cmm}{}

\item{annotation}{}
}
\value{

}
\description{
Title
}
