% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seed.exon.R
\name{.seed.exon.ep.incl}
\alias{.seed.exon.ep.incl}
\title{Seed exons}
\usage{
.seed.exon.ep.incl(cmm = cmm, ep_event = ep_event, keep.intron = keep.intron)
}
\arguments{
\item{cmm}{}

\item{ep_event}{}

\item{keep.intron}{}
}
\value{

}
\description{
Seed exons
}
