% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seed.exon.R
\name{.seed.exon.ip.excl}
\alias{.seed.exon.ip.excl}
\title{Title}
\usage{
.seed.exon.ip.excl(ip_event = ip_event, cmm = cmm, annotation = annotation)
}
\arguments{
\item{ip_event}{}

\item{cmm}{}

\item{annotation}{}
}
\value{

}
\description{
Title
}
