% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPvaluesByContrast.R
\name{.sumPvalsMethod}
\alias{.sumPvalsMethod}
\title{Sum of p-values}
\usage{
.sumPvalsMethod(x, n)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
Internal function of \code{\link{EPrnaseq}}/\code{\link{iPrnaseq}}, not to be called separately.
}
