% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transtruct.prep.R
\name{.transtruct.prep}
\alias{.transtruct.prep}
\title{Transtruct preparation function}
\usage{
.transtruct.prep(
  cmm = cmm,
  event,
  eventlist,
  se = s.exon$se,
  ls_exon = s.exon$ls_exon,
  rs_exon = s.exon$rs_exon,
  eventtype,
  junc.mat = NULL,
  p = p,
  transtruct.ep.incl = NULL,
  transtruct.ep.excl = NULL,
  transtruct.ip.incl = NULL,
  transtruct.ip.excl = NULL,
  s.exon = NULL,
  keep.intron = keep.intron
)
}
\arguments{
\item{cmm}{}

\item{event}{}

\item{eventlist}{}

\item{se}{}

\item{ls_exon}{}

\item{rs_exon}{}

\item{eventtype}{}

\item{junc.mat}{}

\item{p}{}

\item{transtruct.ep.incl}{}

\item{transtruct.ep.excl}{}

\item{transtruct.ip.incl}{}

\item{transtruct.ip.excl}{}

\item{s.exon}{}

\item{keep.intron}{}
}
\value{

}
\description{
Transtruct preparation function
}
