% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPrnaseq.R
\name{EPrnaseq}
\alias{EPrnaseq}
\title{ExonPointer}
\usage{
EPrnaseq(
  Gcount,
  RMM,
  designM,
  contrastM,
  Groups = NULL,
  p = 1,
  threshold = 3,
  annotation = annotation,
  ...
)
}
\arguments{
\item{Gcount}{list; contains gene-wise  matrix of meta-features read counts times samples, generated by \code{\link{countMatrixGenes}}.}

\item{RMM}{gene-wise list that represents the association of exons with other meta-features of genes (introns and junctions (skipping/flanking)). It is generated using \code{\link{readMembershipMatrix}}.}

\item{designM}{design matrix required by limma.}

\item{contrastM}{contrast matrix required by limma.}

\item{Groups}{list of sample groups. \cr
Example: If there are two sample groups with three samples each, 'Groups' should be formed as:
\enumerate{
\item numeric: c(1, 1, 1, 2, 2, 2)
\item alphabetical: c('A', 'A', 'A', 'B', 'B', 'B')}}

\item{p}{number of threads to be used if running in parallel. (default=1)}

\item{threshold}{minimum number of reads that should map to a meta-feature (default=3). If number of reads<threshold, meta-feature would be discarded.}

\item{annotation}{matrix; contains annotation of exons and introns, created using \code{\link{readMembershipMatrix}}.}

\item{...}{other parameters to be passed to \code{\link[limma]{eBayes}}, \code{\link[limma]{voom}}, \code{\link[edgeR]{calcNormFactors}} and \code{\link[limma]{lmFit}} .}
}
\value{
ExonPointer gives a list of ranked cassette exon events with equivalent p-value and t-statistic.
}
\description{
Prediction of cassette exons events by utilizing information of meta-features (flanking junctions, skipping junctions and introns) associated with the exon in context ofs a given gene.
}
\details{
ExonPointer algorithm finds cassette exon events using metafeatures (exons, introns and junctions). The read counts of meta-features are present in Gcount and the association of an exon with introns and junctions (skipping/flanking) is given by Read Membership Matrix (RMM).\cr
    In order to find a cassette exon event, one-tailed p-values of metafeatures are summarized using Irwin-Hall method to find the equivalent P-value (EqP). EqP determines if an event is differentially alternatively spliced. For more details, please refer: S. S. Tabrez, R. D. Sharma, V. Jain, A. A. Siddiqui & A. Mukhopadhyay. Differential alternative splicing coupled to nonsense-mediated decay of mRNA ensures dietary restriction-induced longevity. Nature Communications volume 8, Article number: 306 (2017).
}
\references{
\enumerate{
 \item S. S. Tabrez, R. D. Sharma, V. Jain, A. A. Siddiqui & A. Mukhopadhyay. Differential alternative splicing coupled to nonsense-mediated decay of mRNA ensures dietary restriction-induced longevity. Nature Communications volume 8, Article number: 306 (2017).
 \item Robinson, M. D., McCarthy, D. J. & Smyth, G. K. edgeR: A Bioconductor package for differential expression analysis of digital gene expression data. Bioinformatics 26, 139–140 (2009).
 \item Ritchie, M. E., Phipson, B., Wu, D., Hu, Y., Law, C. W., Shi, W., & Smyth, G. K. limma powers differential expression analyses for RNA-sequencing and microarray studies. Nucleic acids research, 43(7), e47 (2015).
 \item Henrik Bengtsson (2017). matrixStats: Functions that Apply to Rows and Columns of Matrices (and to Vectors). R package version 0.52.2. https://github.com/HenrikBengtsson/matrixStats
 \item https://git.bioconductor.org/packages/Biobase
 \item Huber W, Carey VJ, Gentleman R, Anders S, Carlson M, Carvalho BS, Bravo HC, Davis S, Gatto L, Girke T, Gottardo R, Hahne F, Hansen KD, Irizarry RA, Lawrence M, Love MI, MacDonald J, Obenchain V, Ole's AK, Pag'es H, Reyes A, Shannon P, Smyth GK, Tenenbaum D, Waldron L, Morgan M (2015). “Orchestrating high-throughput genomic analysis with Bioconductor.” Nature Methods, 12(2), 115–121.
 \item https://CRAN.R-project.org/view=HighPerformanceComputing
 }
}
