% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian_snapshot_bs.R
\name{gaussian_snapshot_bs}
\alias{gaussian_snapshot_bs}
\title{Simulate Gaussian edge networks with B-spline latent processes}
\usage{
gaussian_snapshot_bs(n,d,m,self_loops,spline_design,sigma_edge,process_options)
}
\arguments{
\item{n}{A positive integer, the number of nodes.}

\item{d}{A positive integer, the number of latent space dimensions.}

\item{m}{A positive integer, the number of snapshots.
If this argument is not specified, it
is determined from the snapshot time vector \code{spline_design$x_vec}.}

\item{self_loops}{A Boolean, if \code{FALSE}, all diagonal adjacency matrix
entries are set to zero. Defaults to \code{TRUE}.}

\item{spline_design}{A list, describing the \eqn{B}-spline design:
\describe{
\item{q}{A positive integer, the dimension of the \eqn{B}-spline basis.
Must be at least \code{4} and at most \code{m}.}
\item{x_vec}{A vector, the snapshot evaluation times for the data.
Defaults to an equally spaced sequence of length
\code{m} from \code{0} to \code{1}.}
\item{x_max}{A scalar, the maximum of the index space.
Defaults to \code{max(spline_design$x_vec)}.}
\item{x_min}{A scalar, the minimum of the index space.
Defaults to \code{min(spline_design$x_vec)}.}
}}

\item{sigma_edge}{A positive scalar,
the entry-wise standard deviation for the Gaussian edge variables.
Defaults to \code{1}.}

\item{process_options}{A list, containing additional optional arguments:
\describe{
\item{sigma_coord}{A positive scalar, the standard deviation of the
randomly generated basis coordinates. Defaults to \code{1}.}
}}
}
\value{
A list is returned with the realizations of the basis coordinates,
spline design, and the multiplex network snapshots:
\item{A}{An array of dimension \eqn{n \times n \times m}, the realized
functional network data.}
\item{W}{An array of dimension \eqn{n \times q \times d},
the realized basis coordinates.}
\item{spline_design}{A list, describing the \eqn{B}-spline design:
\describe{
\item{type}{The string \code{'bs'}.}
\item{q}{A positive integer, the dimension of the \eqn{B}-spline basis.}
\item{x_vec}{A vector, the snapshot evaluation times for the data.}
\item{x_max}{A scalar, the maximum of the index space.}
\item{x_min}{A scalar, the minimum of the index space.}
\item{spline_matrix}{An \eqn{m \times q} matrix, the B-spline basis
evaluated at the snapshot times.}
}}
}
\description{
\code{gaussian_snapshot_bs} simulates a realization of a functional network
with Gaussian edges, according to an inner product latent process model.
The latent processes are generated from a \eqn{B}-spline basis with equally
spaced knots.
}
\details{
The spline design of the functional network data (snapshot times,
basis dimension) is generated using the information provided in
\code{spline_design}, producing a \eqn{q}-dimensional cubic
\eqn{B}-spline basis with equally spaced knots.

The latent process basis coordinates are generated as iid
Gaussian random variables with standard deviation
\code{process_options$sigma_coord}. Each latent process is given by
\deqn{z_{i,r}(x) = w_{i,r}^{T}B(x).}

Then, the \eqn{n \times n} symmetric adjacency matrix for
snapshot \eqn{k=1,...,m} has independent Gaussian entries
with standard deviation \code{sigma_edge} and mean
\deqn{E([A_k]_{ij}) = z_i(x_k)^{T}z_j(x_k)}
for \eqn{i \leq j} (or \eqn{i < j} with no self loops).
}
