% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getJunctionCountMatrix.R
\name{getJunctionCountMatrix}
\alias{getJunctionCountMatrix}
\title{Generate junction count matrix}
\usage{
getJunctionCountMatrix(files)
}
\arguments{
\item{files}{junction bed files.}
}
\value{
Junction read counts matrix.
}
\description{
This function combines tophat2 pipeline output junctions.bed files after mapping reads to genome/trancriptome. It can be called separately for combining junction.bed files for the FASE pipeline.
}
\references{
\enumerate{
\item  F. Hoffgaard, P. Weil, K. Hamacher. BioPhysConnectoR: Connecting Sequence Information and Biophysical Models. BMC Bioinformatics volume 11, Article number: 199 (2010).
}
}
