% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPvaluesByContrast.R
\name{getPvaluesByContrast}
\alias{getPvaluesByContrast}
\title{Cassette exon/intron retention event ranking by contrast}
\usage{
getPvaluesByContrast(fit, contrast = NULL)
}
\arguments{
\item{fit}{output of addAnnotationRnaSeq function.}

\item{contrast}{contrast whose ranking is required, for example, 'NormalvsTumor' (as used in contrast matrix).}
}
\value{
Data frame that contains ranking of cassette exon/intron retention events in the given contrast or comparision with their annotation. The output can be saved as csv/xlsx file.
}
\description{
Takes the annotated and fitted object of \code{\link{EPrnaseq}}/\code{\link{iPrnaseq}} and the name or number of contrast as given in contrast matrix as input and finds differentially alternatively spliced cassette exon/intron retention events for that contrast.
}
