% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTFnomencJunctionM.R
\name{GTFnomencJunctionM}
\alias{GTFnomencJunctionM}
\title{Junction Matrix Annotation}
\usage{
GTFnomencJunctionM(gtf, JunctionMatrix)
}
\arguments{
\item{gtf}{gtf file of the organism.}

\item{JunctionMatrix}{matrix with read counts of junctions.}
}
\value{
Annotated junction matrix file: JunctionMatix.
}
\description{
This function is used for correcting the annotation of junction matrix in case junction.bed files are obtained from "BAM" files using "regtools" software. This is the case where "BAM" files are downloaded some repository instead of running tophat2 pipeline. This function then produces a correctly annotated junction matrix on the basis of chromosome nomenclature as used in standard "GTF" file.\cr
}
