% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkAnalysis.R
\name{heatscore}
\alias{heatscore}
\title{Heatscore using Splice Index.}
\usage{
heatscore(ep, ip, ep.exp, ip.exp)
}
\arguments{
\item{ep}{ExonPointer file containing cassette exon event ranking.}

\item{ip}{IntronPointer file containing intron retention event ranking.}

\item{ep.exp}{ExonPointer cassete exon expression file.}

\item{ip.exp}{IntronPointer intron retention expression file.}
}
\value{
Heat Score/Splice Index required by HotNet software.
}
\description{
heatscore uses expression of ranked EP and IP events
}
