% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intronGTFparser.R
\name{intronGTFparser}
\alias{intronGTFparser}
\title{intronGTFparser}
\usage{
intronGTFparser(gtf)
}
\arguments{
\item{gtf}{gtf file of the organism.}
}
\value{
gtf file with intron information.
}
\description{
Parse intron location given in a gtf file and updated gtf will be written. Intron information can be used then for counting reads with Rsubread package (check wrapper functions: \code{\link{ppAuto}} and \code{\link{ppSumEIG}} for read count summarization). However, information associated with these introns (related to transcripts) can not be used as annotation since this transcript information is added in the corresponding field to avoid unnecessary errors.
}
\references{
\enumerate{
 \item http://Matrix.R-forge.R-project.org/
 \item Carey V, Long L, Gentleman R (2019). RBGL: An interface to the BOOST graph library. https://bioconductor.org/packages/RBGL/
 \item Gentleman R, Whalen E, Huber W, Falcon S (2019). graph: graph: A package to handle graph data structures. http://www.bioconductor.org/packages/release/bioc/html/graph.html
 }
}
