% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intronMembershipMatrix.R
\name{intronMembershipMatrix}
\alias{intronMembershipMatrix}
\title{Intron Membership Matrix}
\usage{
intronMembershipMatrix(verbose = TRUE, annotation = annotation)
}
\arguments{
\item{verbose}{TRUE}

\item{annotation}{matrix; contains annotation of exons and introns, created using \code{\link{readMembershipMatrix}}.}
}
\value{
intronMembershipMatrix creates gene-wise list which is saved by default as iMM.Rdata. Each gene is represented by a matrix of meta-features times the number of introns in gene. \cr
        A number is assigned for each meta-feature association to introns in the gene as: \cr
        \itemize{
                   \item 0  : No association
                   \item 1  : Exon associated with the intron
                   \item 2  : Intron with itself
                   \item 3  : Junction associated with the intron

        }
}
\description{
iMM describes association of each intron with meta-features (exons, skipping junctions and flanking junctions) of that gene. It can be generated using a gtf file and a combined junction matrix generated via \code{\link{getJunctionCountMatrix}}. iMM is a pre-requisite matrix for running \code{\link{iPrnaseq}}. It should be run only after running \code{\link{readMembershipMatrix}}.
}
\references{
\enumerate{
\item F. Hoffgaard, P. Weil, K. Hamacher. BioPhysConnectoR: Connecting Sequence Information and Biophysical Models. BMC Bioinformatics volume 11, Article number: 199 (2010).
}
}
