% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JunctionMatrixAnnotation.R
\name{JunctionMatrixAnnotation}
\alias{JunctionMatrixAnnotation}
\title{Junction Matrix Annotation}
\usage{
JunctionMatrixAnnotation(gtf, JunctionMatrix)
}
\arguments{
\item{gtf}{gtf file of the organism.}

\item{JunctionMatrix}{matrix containing junction read counts.}
}
\value{
Annotated junction matrix file: JunctionMatixA.
}
\description{
Annotation of junction matrix using gtf file.
}
