% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppAuto.R
\name{ppAuto}
\alias{ppAuto}
\title{RNA Seq and Alternative Splicing preprocessing function}
\usage{
ppAuto(
  folderSRA = FALSE,
  srlist = NULL,
  pairedend = FALSE,
  genomeBI,
  gtf,
  files = "fastq",
  p = 1,
  N = 6,
  r = 44,
  mate_std_dev = 30,
  read_edit_dist = 6,
  max_intron_length = 10000,
  min_intron_length = 50,
  segment_length = NULL,
  ...
)
}
\arguments{
\item{folderSRA}{path of directory containing fastq or SRA files. (default=current directory)}

\item{srlist}{list of unique sample names of fastq/SRA files created by default in the function. Please follow naming convention for the sample files: \cr
For SRA files  : "Sample-S1_1" "Sample-S1_2" (for paired-end reads) and "Sample-S1" (for single-end reads). \cr
For fastq files: "Sample-S1_1.fastq" "Sample-S1_2.fastq" (for paired-end reads) and "Sample-S1.fastq" (for single-end reads). \cr}

\item{pairedend}{boolean, TRUE if reads are paired-end and FALSE if reads are single-end. All files should be either single-end or paired-end. (default=FALSE)}

\item{genomeBI}{path of genome build of the organism created using bowtie2-build command.}

\item{gtf}{intron parsed gtf file of the organism. Please check \code{\link{intronGTFparser}} to generate intron parsed gtf file (to generate intron read counts).}

\item{files}{type of raw read file: fastq or sra (downloaded from NCBI). All files should be in same format and have same read length. (default=fastq)}

\item{p}{number of threads to be utilized by samtools and Rsubread package. (default=1)}

\item{N}{accepted read mismatches. Reads with more than N mismatches are discarded. (default=6) [tophat2 parameter]}

\item{r}{expected inner distance between mate pair. (default=44) [tophat2 parameter]}

\item{mate_std_dev}{the standard deviation for the distribution on inner distances between mate pairs. (default=30) [tophat2 parameter]}

\item{read_edit_dist}{final read alignments having more than these many edit distance are discarded. (default=6) [tophat2 parameter]}

\item{max_intron_length}{when searching for junctions ab initio, TopHat2 will ignore donor/acceptor pairs farther than this many bases apart, except when such a pair is supported by a split segment alignment of a long read. (default=10000) [tophat2 parameter]}

\item{min_intron_length}{topHat2 will ignore donor/acceptor pairs closer than this many bases apart. (default=50) [tophat2 parameter]}

\item{segment_length}{each read is divided into this length and mapped independently to find junctions. [tophat2 parameter]}

\item{...}{other parameter to be passed to tophat2.}
}
\value{
\enumerate{
        \item Mapped, sorted and indexed bam files. (Can be run separately using tophat2 and samtools or wrapper function: \code{\link{ppRawData}})
        \item Lists of gene counts, exon counts and intron counts saved in folderSRA directory as respective Rdata files. (Can be run separately using \code{\link[Rsubread]{featureCounts}} or wrapper function: \code{\link{ppSumEIG}})
        \item Junction Matrix: Matrix with annotated junction count reads. (Can be run separately using \code{\link{getJunctionCountMatrix}} or wrapper function: \code{\link{ppRawData}})
        \item RMM            : ReadMembershipMatrix. (Can be run separately using \code{\link{readMembershipMatrix}} or wrapper function: \code{\link{ppFASE}})
        \item iMM            : intronMembershipMatrix. (Can be run separately using \code{\link{intronMembershipMatrix}} or wrapper function: \code{\link{ppFASE}})
        \item Gcount         : A list of gene-wise read count summarization of meta-features times samples in the study. (Can be run separately using \code{\link{countMatrixGenes}} or wrapper function: \code{\link{ppFASE}})
        }
}
\description{
ppAuto is a wrapper function for several tools and functions that perform preprocessing of RNA Sequencing data. This function performs preprocessing that includes mapping of reads, sorting and indexing of bam files, to summarization of read counts for exons, introns, genes and junctions. ppAuto also creates several prerequisite matrices including junction matrix, ReadMembershipMatrix (RMM), IntronMembershipMatrix (iMM) and Gcount matrix in order to run ExonPointer and IntronPointer algorithms. \cr

System requirements for ppAuto include: \cr
    \enumerate{
              \item fastq-dump (if files='SRA')
              \item tophat2
              \item samtools
}
}
\references{
\enumerate{
\item Liao Y, Smyth GK, Shi W. The R package Rsubread is easier, faster, cheaper and better for alignment and quantification of RNA sequencing reads. Nucleic Acids Research, 47, e47 (2019).
}
}
