% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppSumEIG.R
\name{ppSumEIG}
\alias{ppSumEIG}
\title{RNA Seq Preprocessing Read Summarization}
\usage{
ppSumEIG(
  folderSRA = FALSE,
  pairedend = FALSE,
  p = 1,
  gtf = gtf,
  srlist = NULL,
  ...
)
}
\arguments{
\item{folderSRA}{path of directory containing aligned and indexed bam file folders. (default=current directory)}

\item{pairedend}{boolean, TRUE if reads are paired-end and FALSE if reads are single-end. (default=FALSE).}

\item{p}{number of threads to be utilized by Rsubread package. (default=1)}

\item{gtf}{intron parsed gtf file of the organism.}

\item{...}{other parameters to be passed to \code{\link[Rsubread]{featureCounts}}.}
}
\value{
Lists of gene counts, exon counts and intron counts saved in folderSRA directory as respective Rdata files. (Can be run separately using \code{\link[Rsubread]{featureCounts}} or automatic wrapper function for entire pre-processing: \code{\link{ppAuto}})
}
\description{
ppSumEIG is a manual wrapper function that provides summarization of read counts for exons, introns and genes using \code{\link[Rsubread]{featureCounts}}. The gtf file passed to this function should first be passed to \code{\link{intronGTFparser}} to find the location of introns. Reads used for summarization by ppSumEIG should already be mapped, sorted and index using tophat2 and samtools or their wrapper function: \code{\link{ppRawData}}. The summarized counts produced by ppSumEIG can be further processed using \code{\link{ppFASE}}, which produces several matrices required by ExonPointer and IntronPointer algorithms for finding alternative splicing events. ppSumEIG need not be run if \code{\link{ppAuto}} has already been run.
}
\references{
\enumerate{
\item Liao Y, Smyth GK, Shi W. The R package Rsubread is easier, faster, cheaper and better for alignment and quantification of RNA sequencing reads. Nucleic Acids Research, 47, e47 (2019).
}
}
