% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{proc_align_slicewise3}
\alias{proc_align_slicewise3}
\title{Slicewise Procrustes alignment for 3-mode tensors}
\usage{
proc_align_slicewise3(A,B)
}
\arguments{
\item{A}{An \eqn{n \times d \times m} array.}

\item{B}{An \eqn{n \times d \times m} array.}
}
\value{
Returns the \eqn{n \times d \times m}
array resulting from applying the optimal aligning transformations to
the columns of the \eqn{n \times d} slices of \code{A}.
}
\description{
\code{proc_align_slicewise3} applies an orthogonal transformation
to the columns of each of the \eqn{n \times d} slices of an
\eqn{n \times d \times m} array \eqn{A} to
find the best approximation (in terms of matrix Frobenius norm) to
the corresponding \eqn{n \times d} slice of a
second \eqn{n \times d \times m} array \eqn{B}.
}
