% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMembershipMatrix.R
\name{readMembershipMatrix}
\alias{readMembershipMatrix}
\title{Read Membership Matrix}
\usage{
readMembershipMatrix(gtf, JunctionMatrix = JunctionMatrix)
}
\arguments{
\item{gtf}{gtf file of the organism.}

\item{JunctionMatrix}{junction matrix contains read counts of each junction mapped by tophat2 alongwith their annotation.}
}
\value{
readMembershipMatrix creates a gene-wise list which is saved by default as RMM.Rdata. Each gene is represented by a matrix of meta-features times the number of exons in gene. \cr
        A number is assigned for each meta-feature association to exons in the gene as: \cr
        \itemize{
                   \item 0  : No association
                   \item 0.5: Skipping junction to the exon
                   \item 1  : Exon with itself
                   \item 2  : Flanking junction to the exon
                   \item 3  : Intron associated with the exon

        }
}
\description{
RMM describes association of each exon with meta-features (introns, skipping junctions and flanking junctions) of that gene. It can be generated using a gtf file and a combined junction matrix generated via \code{\link{getJunctionCountMatrix}}. RMM is a pre-requisite matrix for running \code{\link{EPrnaseq}}.
}
\references{
\enumerate{
\item F. Hoffgaard, P. Weil, K. Hamacher. BioPhysConnectoR: Connecting Sequence Information and Biophysical Models. BMC Bioinformatics volume 11, Article number: 199 (2010).
}
}
