% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkAnalysis.R
\name{simscore}
\alias{simscore}
\title{Similarity Score for network edge weight}
\usage{
simscore(ep, ep.exp, ip, ip.exp)
}
\arguments{
\item{ep}{ExonPointer file containing cassette exon event ranking.}

\item{ep.exp}{ExonPointer cassete exon expression file.}

\item{ip}{IntronPointer file containing intron retention event ranking.}

\item{ip.exp}{IntronPointer intron retention expression file.}
}
\value{
Similarity Score gives cosine similarity between the splice index of two genes. It can be used as edge weight for an unweighted network.
}
\description{
Similarity Score for network edge weight
}
