% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survFASE.R
\name{survFASE}
\alias{survFASE}
\title{Survival analysis using metafeatures}
\usage{
survFASE(
  Time = Time,
  Status = Status,
  Gcount,
  clinical.data,
  rmm = NULL,
  imm = NULL,
  eventID,
  threshold = 0.6,
  designM,
  Groups
)
}
\arguments{
\item{Time}{column name of survival time in clinical data. (default = Time)}

\item{Status}{column name of survival status (alive/dead status) in clinical data. 1 = dead and 0 = alive. (default = Status)}

\item{Gcount}{Gene count matrix for the given gene. It should contain raw meta-feature counts of only patients. Rownames of gcount should be unique sample IDs that can be mapped to the clinical data.}

\item{clinical.data}{clinical data of patients. It should contain survival time (days to last follow up) and survival status (0/1) of the patients. Rownames of clinical data should be unique, non-repeating sample IDs that can be mapped to the Gcount.}

\item{rmm}{readMembershipMatrix matrix of the gene. It contains association between exons and other metafeatures in a gene and is generated by default as RMM.Rdata using \code{\link{readMembershipMatrix}} function. RMM is required only for finding survival rate associated with a cassette exon event.}

\item{imm}{intronMembershipMatrix matrix of the gene. It contains association between introns and other metafeatures in a gene. It is generated by default as iMM.Rdata using \code{\link{intronMembershipMatrix}} function. iMM is required only for finding survival rate associated with an intron retention event.}

\item{eventID}{exonID or intronID of the AS event for survival analysis.}

\item{designM}{design matrix}

\item{Groups}{list of sample groups}
}
\value{
survFASE returns an overall p-value, concordance index and Cox-PH statistics. The overall p-value suggests whether or not the given exon/intron significantly affects patient survival. C-index signifies goodness-of-fit of the model. Cox-PH results show which of the metafeatures associated with the exon/intron affect survival rate and their statistical inferences like hazard-ratio, beta-coefficient, etc.
}
\description{
survFASE finds survival rate of patients using altenative splicing data. It requires exon, intron and junction expression of the samples generated using FASE (check \code{\link{EPrnaseq}}, \code{\link{iPrnaseq}} and \code{\link{DEJ}}). rownames of all expression files and clinical data should have the same identifier/sample ID, otherwise survFASE would not be able to perform survival analysis. survFASE uses RMM/iMM to find metafeature(s) associated with the given exonID/intronID and incorporate the expression of those metafeatures with respective exonID/intronID.
}
