% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transconc.R
\name{transconc}
\alias{transconc}
\title{Transcript Concentration}
\usage{
transconc(transtruct, designM)
}
\arguments{
\item{transtruct}{transcript structure for EP/IP inclusion and exclusion for a gene.}

\item{designM}{design matrix for the two conditions in question. It should be a binary matrix where 1 represents samples of the condition and 0 represents the samples of other condition.}
}
\value{
transconc returns three matrices: \cr
\enumerate{
         \item transconc.samples: transcript concentration by samples.
         \item transconc.condition: transcript concentration by condition.
         \item transconc.TS: structures of transcripts and concentration of their meta-features, used for evaluating transcript abundance (transconc combines all structures by transtruct and retains only unique structures).
}
}
\description{
Transcript concentration gives the abundance of transcripts expressed in different samples and conditions. It uses transcript structure generated by transtruct method and superimposes expression values of corresponding meta-features.
}
