% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survFASE.R
\name{survFASE}
\alias{survFASE}
\title{Survival analysis using metafeatures}
\usage{
survFASE(
  junction.expression,
  ep.expression = NULL,
  ip.expression = NULL,
  clinical.data,
  rmm = NULL,
  imm = NULL,
  exonID = NULL,
  intronID = NULL,
  threshold = 0.6
)
}
\arguments{
\item{junction.expression}{log2CPM expression of junctions generated by \code{\link{DEJ}} function (samples*junctions). Before passing this file to survFASE, expression of normal/control samples should be removed, as survFASE requires only treated/patient data. Rownames of junction expression should be unique sample IDs that can be mapped to the clinical data and column names should be unique junction IDs..}

\item{ep.expression}{log2CPM expression of exons (samples*exons). This file should contain exon expression of only patients/treated samples and expression of control/normal samples should be removed. Rownames of exon expression should be unique sample IDs that can be mapped to the clinical data.}

\item{ip.expression}{log2CPM expression of introns (samples*introns). This file should contain intron expression of only patients/treated samples and expression of control/normal samples should be removed. Rownames of intron expression should be unique sample IDs that can be mapped to the clinical data.}

\item{clinical.data}{clinical data of patients. It should contain survival time (days to last follow up) and survival status (0/1) of the patients. Rownames of clinical data should be unique sample IDs that can be mapped to the expression data.}

\item{rmm}{readMembershipMatrix matrix of the gene. It contains association between exons and other metafeatures in a gene and is generated by default as RMM.Rdata using \code{\link{readMembershipMatrix}} function.}

\item{imm}{intronMembershipMatrix matrix of the gene. It contains association between introns and other metafeatures in a gene. It is generated by default as iMM.Rdata using \code{\link{intronMembershipMatrix}} function. iMM is required only for finding survival rate associated with an intron retention event.}

\item{exonID}{exonID for survival analysis.}

\item{intronID}{intronID for survival analysis.}
}
\value{
survFASE returns an overall p-value, concordance index and Cox-PH statistics. The overall p-value suggests whether or not the given exon/intron significantly affects patient survival. C-index signifies goodness-of-fit of the model. Cox-PH results show which of the metafeatures associated with the exon/intron affect survival rate and their statistical inferences like hazard-ratio, beta-coefficient, etc.
}
\description{
survFASE finds survival rate of patients using altenative splicing data. It requires exon, intron and junction expression of the samples generated using FASE (check \code{\link{EPrnaseq}}, \code{\link{iPrnaseq}} and \code{\link{DEJ}}). rownames of all expression files and clinical data should have the same identifier/sample ID, otherwise survFASE would not be able to perform survival analysis. survFASE uses RMM/iMM to find metafeature(s) associated with the given exonID/intronID and incorporate the expression of those metafeatures with respective exonID/intronID.
}
